inference/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# SSE support
sse-starlette==2.0.0

# HTTP client for consuming SSE
httpx==0.27.0

# Data validation
pydantic==2.6.3

# MongoDB
pymongo[srv]==4.6.1

# Environment variables
python-dotenv==1.0.0

# Fireworks.ai (OpenAI-compatible API)
openai==1.12.0

# For face recognition (optional)
# dlib==19.24.2
# face-recognition==1.3.0

# For numerical computations
numpy==1.24.3

# For image processing (optional)
# Pillow==10.2.0